[frontend] Migrate the TopMenuData to typescript (#2535)
